move NameIdentifierUtil to the util sub module
