+++ core/src/main/java/com/datastrato/gravitino/utils/NameIdentifierUtil.java
@@ -2,8 +2,10 @@
  * Copyright 2023 Datastrato Pvt Ltd.
  * This software is licensed under the Apache License version 2.
  */
-package com.datastrato.gravitino;
+package com.datastrato.gravitino.utils;
 
+import com.datastrato.gravitino.NameIdentifier;
+import com.datastrato.gravitino.Namespace;
 import com.datastrato.gravitino.exceptions.IllegalNameIdentifierException;
 
 /**
