+++ core/src/test/java/com/datastrato/gravitino/catalog/TestPartitionOperationDispatcher.java
@@ -5,7 +5,6 @@
 package com.datastrato.gravitino.catalog;
 
 import com.datastrato.gravitino.NameIdentifier;
-import com.datastrato.gravitino.NameIdentifierUtil;
 import com.datastrato.gravitino.rel.Column;
 import com.datastrato.gravitino.rel.expressions.literals.Literal;
 import com.datastrato.gravitino.rel.expressions.literals.Literals;
@@ -13,6 +12,7 @@ import com.datastrato.gravitino.rel.partitions.Partition;
 import com.datastrato.gravitino.rel.partitions.Partitions;
 import com.datastrato.gravitino.rel.types.Types;
 import com.datastrato.gravitino.utils.IsolatedClassLoader;
+import com.datastrato.gravitino.utils.NameIdentifierUtil;
 import com.google.common.collect.Maps;
 import java.util.Arrays;
 import org.junit.jupiter.api.Assertions;
