+++ server/src/main/java/com/datastrato/gravitino/server/web/rest/TableOperations.java
@@ -10,7 +10,6 @@ import static com.datastrato.gravitino.dto.util.DTOConverters.fromDTOs;
 import com.codahale.metrics.annotation.ResponseMetered;
 import com.codahale.metrics.annotation.Timed;
 import com.datastrato.gravitino.NameIdentifier;
-import com.datastrato.gravitino.NameIdentifierUtil;
 import com.datastrato.gravitino.Namespace;
 import com.datastrato.gravitino.catalog.TableDispatcher;
 import com.datastrato.gravitino.dto.requests.TableCreateRequest;
@@ -26,6 +25,7 @@ import com.datastrato.gravitino.metrics.MetricNames;
 import com.datastrato.gravitino.rel.Table;
 import com.datastrato.gravitino.rel.TableChange;
 import com.datastrato.gravitino.server.web.Utils;
+import com.datastrato.gravitino.utils.NameIdentifierUtil;
 import javax.inject.Inject;
 import javax.servlet.http.HttpServletRequest;
 import javax.ws.rs.DELETE;
